Improve users getter code
